+++ src/app/page.tsx
@@ -1,13 +1,11 @@
 import GameApp from "@/components/customUi/GameApp";
-import SurveyDrawerDemo from "./survey/survey-action/Survey";
-import { Alert } from "@/components/ui/alert";
-import { AlertDemo } from "@/components/customUi/AlertDemo";
+import SurveyDrawerBtn from "./survey/survey-action/SurveyDrawerBtn";
 
 export default async function Home() {
   return (
     <>
       <GameApp />
-      <SurveyDrawerDemo />
+      <SurveyDrawerBtn />
     </>
   );
 }
